# UI/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TF26 API",
    "description": "Turkish participation-banking campaigns and live pricing. This service authenticates callers and exposes what `banks/`, `index/` and `corpus/` already do; it holds no banking logic itself.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/compare-tables": {
      "get": {
        "tags": [
          "compare-tables"
        ],
        "summary": "List Tables",
        "description": "Every table in one category, for the browse-by-subcategory picker.",
        "operationId": "list_tables",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "'ürün' or 'kampanya'.",
              "title": "Category"
            },
            "description": "'ürün' or 'kampanya'."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/compare-tables/{table_id}": {
      "get": {
        "tags": [
          "compare-tables"
        ],
        "summary": "Get Table",
        "description": "One table, shaped for `<ProducedTable />`.\n\nColumn order is: a `Banka` column first (bank keys the frontend already\nknows how to render), then every column the table itself declared in the\norder the producer chose, then the two derived validity columns.\n\nThe validity pair is split rather than shown as one range string because the\ntwo halves are read differently: the verdict is a `badge`, which\n`resolveTable` makes filterable, and the end date is a `date`, which it makes\nsortable. One combined column would be neither — and filtering by whether an\noffer is still live, and sorting by when it ends, are the two things this\ndata is for.",
        "operationId": "get_table",
        "parameters": [
          {
            "name": "table_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "An id from GET /api/compare-tables.",
              "title": "Table Id"
            },
            "description": "An id from GET /api/compare-tables."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/compare-tables/{table_id}/overview": {
      "get": {
        "tags": [
          "compare-tables"
        ],
        "summary": "Get Overview",
        "description": "The overview, or what is happening instead.\n\nDeliberately not \"generate one if it is missing\". Writing an overview needs\nthe page as the browser sees it — an outline and a screenshot this endpoint\nhas no way to produce — so the client asks here first and only pays for a\ncapture when the answer is `missing`. A GET that quietly cost a\nvision-model call would also be a GET that is not safe to retry.\n\n`generating` is the reason this returns a wrapper instead of a 404. A model\nreading a screenshot takes a minute on a quiet host and several on a busy\none, and a client that cannot tell \"still working\" from \"nothing here\" has\nto guess a timeout — which is either too short for the busy host or leaves\na card spinning at a dead one.",
        "operationId": "get_overview",
        "parameters": [
          {
            "name": "table_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "An id from GET /api/compare-tables.",
              "title": "Table Id"
            },
            "description": "An id from GET /api/compare-tables."
          },
          {
            "name": "locale",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "'tr' or 'en'.",
              "default": "tr",
              "title": "Locale"
            },
            "description": "'tr' or 'en'."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableOverviewState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "compare-tables"
        ],
        "summary": "Create Overview",
        "description": "Start writing the overview for one table. Poll the GET for the result.\n\n**Accepted, not written.** A generation takes 70-120 seconds, which is\nlonger than the things between this and the browser are willing to hold a\nsocket open -- the dev server's `/api` proxy cuts it at 30. So the work\nstarts in a thread and the client polls the GET it already calls on\narrival, which is the same row every later visitor will read.\n\nThe table the agent reads is `get_table`'s own output, not the payload the\ncaller sent: the browser supplies what it *saw* -- the outline and the\nscreenshot -- and the figures come from the same function that rendered\nthem, so nothing on the wire can talk the agent into summarising a table\nthe server does not have.\n\nNo authentication, like the rest of this router: the pool is public data.\nThe cost of that is bounded by the cache, which answers every repeat for\nfree, and by the single-flight lock, which collapses a stampede on one\ntable into one call. If it ever needs a gate, it needs a rate limit rather\nthan a login.",
        "operationId": "create_overview",
        "parameters": [
          {
            "name": "table_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "An id from GET /api/compare-tables.",
              "title": "Table Id"
            },
            "description": "An id from GET /api/compare-tables."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TableOverviewRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableOverviewStarted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ColumnOut": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label"
        ],
        "title": "ColumnOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PageContext": {
        "properties": {
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text",
            "description": "The page as a semantic outline, in markdown, wrapped in a `<page-snapshot>` element. Deliberately unbounded, like every other payload on its way to a model in this app: half a page answers a question about the other half wrongly."
          }
        },
        "type": "object",
        "title": "PageContext",
        "description": "What the browser saw, for the overview agent to read.\n\nText only. A screenshot used to travel here too, and it was dropped: it cost\nminutes of vision prefill per table and carried nothing the outline does not\n— once `data-outline-list` taught the outline to keep short-line cards like\n\"banks that do not offer this\", which is the one thing the picture had been\ncovering for."
      },
...
      "RankedBankOut": {
        "properties": {
          "bank": {
            "type": "string",
            "title": "Bank"
          },
          "why": {
            "type": "string",
            "title": "Why"
          }
        },
        "type": "object",
        "required": [
          "bank",
          "why"
        ],
        "title": "RankedBankOut"
      },
...
      "RowOut": {
        "properties": {
          "cells": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Cells"
          },
          "cite_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cite Url"
          },
          "cite_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cite Note",
            "description": "The pipeline's own note on why this source supports the row — shown as a hover title on the citation link, not a second dashboard."
          },
          "offers": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offers",
            "description": "Whether this bank offers the thing at all. Read here rather than in the browser because the producer answers it in a column it names itself — 93 different names across the pool, and in some tables no column at all. `null` means nothing in the row settles it, which is not the same as `false`: an unclassified row stays in the table."
          },
          "cell_notes": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Cell Notes",
            "description": "Column key -> a note to hover on that one cell. Carries the full validity window behind the verdict chip, so a table already running to 22 columns does not need a second date column to show a start date."
          },
          "cell_tones": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Cell Tones",
            "description": "Column key -> the tone its badge cell carries ('neutral' | 'ok' | 'warn' | 'bad'). Sent as data because only this layer knows which of its own values is the good one; the table renderer draws badges for every producer and cannot infer it from the string."
          }
        },
        "type": "object",
        "required": [
          "cells"
        ],
        "title": "RowOut"
      },
...
      "TableDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "subtitle": {
            "type": "string",
            "title": "Subtitle",
            "default": ""
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnOut"
            },
            "type": "array",
            "title": "Columns"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/RowOut"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "columns",
          "rows"
        ],
        "title": "TableDetailOut",
        "description": "Shaped exactly like the frontend's `TableProps` (`UI/src/lib/contract.ts`)\nso it can be spread straight into `<TableWidget />` with no translation\nlayer on the client."
      },
      "TableListOut": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "subcategories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Subcategories"
          },
          "tables": {
            "items": {
              "$ref": "#/components/schemas/TableSummaryOut"
            },
            "type": "array",
            "title": "Tables"
          }
        },
        "type": "object",
        "required": [
          "category",
          "subcategories",
          "tables"
        ],
        "title": "TableListOut",
        "description": "Every table in one category, plus the subcategories present in it.\n\n`subcategories` is derived from the tables actually returned, not the\nglobal registry — a category-scoped dropdown should never offer a\nsubcategory that has zero tables under this category."
      },
...
      "TableOverviewOut": {
        "properties": {
          "table_id": {
            "type": "string",
            "title": "Table Id"
          },
          "locale": {
            "type": "string",
            "title": "Locale"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "recommended": {
            "items": {
              "$ref": "#/components/schemas/RankedBankOut"
            },
            "type": "array",
            "title": "Recommended",
            "description": "At most two, best first. Empty when the page supports no pick."
          },
          "not_recommended": {
            "items": {
              "$ref": "#/components/schemas/RankedBankOut"
            },
            "type": "array",
            "title": "Not Recommended",
            "description": "At most two: weakest terms, or banks that do not offer this."
          },
          "caveat": {
            "type": "string",
            "title": "Caveat",
            "default": ""
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "table_id",
          "locale",
          "summary",
          "generated_at"
        ],
        "title": "TableOverviewOut",
        "description": "One table, as the overview agent read it."
      },
      "TableOverviewRequest": {
        "properties": {
          "locale": {
            "type": "string",
            "title": "Locale",
            "description": "'tr' or 'en'. Anything else is Turkish.",
            "default": "tr"
          },
          "page": {
            "$ref": "#/components/schemas/PageContext"
          }
        },
        "type": "object",
        "title": "TableOverviewRequest",
        "description": "Ask for an overview of one table, offering the page as evidence."
      },
      "TableOverviewStarted": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "generating",
              "already_running"
            ],
            "title": "Status",
            "description": "`already_running` means another reader asked first and the same row will answer both. Either way the client polls the GET."
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "TableOverviewStarted",
        "description": "The answer to \"please write one\": it is being written."
      },
      "TableOverviewState": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "generating",
              "missing"
            ],
            "title": "Status"
          },
          "overview": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TableOverviewOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "TableOverviewState",
        "description": "Whether there is an overview for this table, and if not, why not.\n\nA wrapper rather than \"200 or 404\", because \"nobody has asked for one\" and\n\"one is being written right now\" are different answers and the client acts\non each differently: the first means start one, the second means keep\nwaiting. Told apart by a 404 alone, a card either gives up on a slow model\nor spins forever on a dead one."
      },
      "TableSummaryOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "docstring": {
            "type": "string",
            "title": "Docstring"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "description": "'ürün' or 'kampanya'."
          },
          "subcategory": {
            "type": "string",
            "title": "Subcategory"
          }
        },
        "type": "object",
        "required": [
          "id",
          "topic",
          "docstring",
          "category",
          "subcategory"
        ],
        "title": "TableSummaryOut",
        "description": "One table, without its rows — enough to list and pick from."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}